override renderData method in 1D Array tracer to avoid special case for heapsort, bst in 2D tracer
